[Wiki] Force length API param to int

+++ commands/Wikipedia.py
@@ -53,7 +53,7 @@ class Command(CommandTemplate):
 				  'exintro': '1', 'explaintext': '1', 'exsectionformat': 'plain', 'titles': pagename}
 		#If we need to be verbose, get as many characters as we can
 		if addExtendedText:
-			params['exchars'] = Constants.MAX_MESSAGE_LENGTH * 1.2  #Get some extra characters since we may remove some later
+			params['exchars'] = int(Constants.MAX_MESSAGE_LENGTH * 1.2)  #Get some extra characters since we may remove some later
 		#Otherwise just get the first sentence
 		else:
 			params['exsentences'] = '1'
